Immoscraper 1.1
After an extensive overhaul, two new variables are added: "Immobilie" and "Typ".

"Immobilie" indicates whether the real estate property is a flat or a house. "Typ" indicates the type of ownership. This means that property is either for rent or for sale. This increases the data scale by large margins and should make room for better insights.

In further updates, I plan on including additional crawlers for land and for commercial real estate.